move compressors out into their own files
